fixed open_gui
